- Fixed removing tmp dir when it didn't set.

+++ LibHooks/tests/testsuite/PreCommit/Command/InstallTest.php
@@ -84,7 +84,9 @@ BODY;
      */
     protected function tearDown()
     {
-        `rm -rf {$this->_tmp}/*`;
+        if (is_dir($this->_tmp)) {
+            `rm -rf {$this->_tmp}/*`;
+        }
         parent::tearDown();
     }
 }
